tests/wpt/tests/css/css-grid/masonry/tentative/track-sizing/auto-repeat/row-auto-repeat-011.html

<!DOCTYPE html>
<title>Auto repeat tracks with min and max sizes</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<meta name="assert" content="This test checks that auto repeat tracks don't overflow the masonry container size when max size is definite, even if min size is bigger than that.">
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<style>
.masonry {
    position: relative;
    display: masonry;
    grid-template-rows: repeat(auto-fill, 50px);
    masonry-direction: row;
    max-width: 100px;
    min-width: 250px;
    max-height: 50px;
    min-height: 125px;
    float: left;
    background: pink;
}
.item {
  background: lime;
  /* Place item on the last row. */
  grid-row: -2;
  width: 250px;
  height: 100%;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.masonry');">

<div id="log"></div>

<div class="masonry" data-expected-width="250" data-expected-height="125">
  <div class="item" data-offset-x="0" data-offset-y="50" data-expected-width="250" data-expected-height="50"></div>
</div>

<div class="masonry" style="width: 200px; height: 100px;" data-expected-width="250" data-expected-height="125">
  <div class="item" data-offset-x="0" data-offset-y="50" data-expected-width="250" data-expected-height="50"></div>
</div>

<div class="masonry" style="width: min-content; height: min-content;" data-expected-width="250" data-expected-height="125">
  <div class="item" data-offset-x="0" data-offset-y="50" data-expected-width="250" data-expected-height="50"></div>
</div>

<div class="masonry" style="width: max-content; height: max-content;" data-expected-width="250" data-expected-height="125">
  <div class="item" data-offset-x="0" data-offset-y="50" data-expected-width="250" data-expected-height="50"></div>
</div>

</body>
